+++ Content.Server/IoC/ServerContentIoC.cs
@@ -1,5 +1,6 @@
 using Content.Server._Sunrise.Contributors;
 using Content.Server._Sunrise.IoC;
+using Content.Server._Sunrise.PlanetPrison;
 using Content.Server._Sunrise.PlayerCache;
 using Content.Server._Sunrise.ServersHub;
 using Content.Server._Sunrise.TTS;
@@ -91,6 +92,7 @@ internal static class ServerContentIoC
         // Sunrise-Start
         deps.Register<ServersHubManager>();
         deps.Register<ContributorsManager>();
+        deps.Register<PlanetPrisonMapManager>();
         deps.Register<PlayerCacheManager>();
         deps.Register<TTSManager>();
         deps.Register<IIPBlockingSystem, IPBlockingSystem>();
